+++ src/3_game_loop/a_game_logic/entities/blight_entity/blight_collision.c
@@ -6,7 +6,7 @@
 /*   By: fschuber <fschuber@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/26 21:57:58 by freddy            #+#    #+#             */
-/*   Updated: 2024/07/03 12:16:28 by fschuber         ###   ########.fr       */
+/*   Updated: 2024/07/04 11:32:18 by fschuber         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,6 +26,7 @@ void	on_collision_blight(t_entity *self, t_entity *other)
 	else
 		health_decrease_amount = MINOR_PROJECTILE_HIT_DECREASE;
 	self->health -= health_decrease_amount;
+	blight->hurt_state = BLIGHT_HURT_STATE_FRAME_DURATION;
 	if (self->health > 0)
 		logger_v(LOGGER_ACTION, "Blight hit by projectile!");
 	else